Add trait for dates handling
