+++ src/Core/DateTrait.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace App\Core;
+
+trait DateTrait
+{
+    /** @var \DateTime */
+    private $createdAt;
+    /** @var \DateTime */
+    private $updatedAt;
+
+    public function getCreatedAt(): \DateTime
+    {
+        return $this->createdAt;
+    }
+
+    public function setCreatedAt(\DateTime $createdAt): void
+    {
+        $this->createdAt = $createdAt;
+    }
+
+    public function getUpdatedAt(): \DateTime
+    {
+        return $this->updatedAt;
+    }
+
+    public function setUpdatedAt(\DateTime $updatedAt): void
+    {
+        $this->updatedAt = $updatedAt;
+    }
+}
